Move signature methods into separate signatures mod
